rename current vpn to legacy
