+++ network-protection/network-protection-impl/src/main/res/layout/legacy_view_settings_netp.xml
